Update from .NET Framework 4.6 to .NET 7 (#3729)
Updates from net46->net7, dropping support for mono in favor of using the latest runtime (along with the performance/API improvements). Releases will be posted as 64bit only for now.

Refactors a good amount of internal API methods to be more performant and more customizable for future updates & fixes.

Adds functionality for Batch Editor commands to `>`, `<` and <=/>=

TID/SID properties renamed to TID16/SID16 for clarity; other properties exposed for Gen7 / display variants.

Main window has a new layout to account for DPI scaling (8 point grid)

Fixed: Tatsugiri and Paldean Tauros now output Showdown form names as Showdown expects
Changed: Gen9 species now interact based on the confirmed National Dex IDs (closes #3724)
Fixed: Pokedex set all no longer clears species with unavailable non-base forms (closes #3720)
Changed: Hyper Training suggestions now apply for level 50 in SV. (closes #3714)
Fixed: B2/W2 hatched egg met locations exclusive to specific versions are now explicitly checked (closes #3691)
Added: Properties for ribbon/mark count (closes #3659)
Fixed: Traded SV eggs are now checked correctly (closes #3692)

+++ PKHeX.Core/Legality/Learnset/LearnsetReader.cs
@@ -15,7 +15,7 @@ public static class LearnsetReader
     /// </summary>
     /// <param name="input">Raw ROM data containing the contiguous moves</param>
     /// <param name="maxSpecies">Highest species ID for the input game.</param>
-    public static Learnset[] GetArray(ReadOnlySpan<byte> input, int maxSpecies)
+    public static Learnset[] GetArray(ReadOnlySpan<byte> input, ushort maxSpecies)
     {
         int offset = 0;
         var result = new Learnset[maxSpecies + 1];
