+++ PKHeX.Core/PersonalInfo/Interfaces/IPersonalInfo.cs
@@ -13,7 +13,7 @@ public interface IPersonalInfo : IBaseStat, IEffortValueYield, IGenderDetail, IP
     /// <summary>
     /// Gender Ratio value determining if the entry is a fixed gender or bi-gendered.
     /// </summary>
-    int Gender { get; set; }
+    byte Gender { get; set; }
 
     /// <summary>
     /// Experience-Level Growth Rate type
