+++ PKHeX.Core/PersonalInfo/Table/PersonalTable8BDSP.cs
@@ -29,12 +29,12 @@ public sealed class PersonalTable8BDSP : IPersonalTable, IPersonalTable<Personal
 
     public int GetFormIndex(ushort species, byte form)
     {
-        if ((uint)species <= MaxSpecies)
+        if (species <= MaxSpecies)
             return Table[species].FormIndex(species, form);
         return 0;
     }
 
-    public bool IsSpeciesInGame(ushort species) => (uint)species <= MaxSpecies;
+    public bool IsSpeciesInGame(ushort species) => species <= MaxSpecies;
     public bool IsPresentInGame(ushort species, byte form)
     {
         if (!IsSpeciesInGame(species))
